Revert "Refactor code (#4)"
This reverts commit 14f14698dc92198064d57b115aa25bc9985fce3d.
