+++ src/scenes/GameScene.ts
@@ -1,34 +1,33 @@
 import Phaser from "phaser";
 import Player from "../objects/Player";
 import TrickArea from "../objects/TrickArea";
+import GameManager from "../managers/GameManager";
 import AudioManager from "../managers/AudioManager";
 import NetworkIndicator from "../components/NetworkIndicator";
+import { EVENTS, type Suit } from "../utils/constants";
 import { ReconnectionOverlay } from "../objects/game/ReconnectionOverlay";
+import NetworkManager from "../managers/NetworkManager";
+import type { CardData } from "../type";
 import Common from "../objects/game/Common";
-import GameModeFactory from "../modes/GameModeFactory";
-import type { GameModeBase } from "../modes/GameModeBase";
-import { EVENTS } from "../utils/constants";
 
 export default class GameScene extends Phaser.Scene {
-  private gameMode!: GameModeBase;
+  isMultiplayer: boolean;
+  networkManager!: NetworkManager;
   audioManager!: AudioManager;
+  gameManager!: GameManager;
   trickArea!: TrickArea;
   players!: Player[];
-  networkIndicator?: NetworkIndicator;
-  reconnectionOverlay?: ReconnectionOverlay;
-  private initData: any;
+  networkIndicator!: NetworkIndicator;
+  reconnectionOverlay!: ReconnectionOverlay;
 
   constructor() {
     super({ key: "GameScene" });
+    this.isMultiplayer = false;
   }
 
   init(data: any) {
-    // Store init data for later use
-    this.initData = data;
-
-    // Create game mode using factory
-    const modeType = GameModeFactory.getTypeFromInitData(data);
-    this.gameMode = GameModeFactory.createGameMode(modeType);
+    this.isMultiplayer = data?.isMultiplayer || false;
+    this.networkManager = data?.networkManager || null;
   }
 
   create() {
@@ -50,77 +49,168 @@ export default class GameScene extends Phaser.Scene {
     // Create trick area
     this.trickArea = new TrickArea(this);
 
-    // Initialize game mode
-    this.initializeGameMode();
+    if (this.isMultiplayer) {
+      this.setupMultiplayer();
+    } else {
+      this.setupSoloGame();
+    }
   }
 
-  async initializeGameMode() {
-    // Initialize the game mode with all dependencies
-    await this.gameMode.initialize(this, {
-      trickArea: this.trickArea,
-      audioManager: this.audioManager,
-      networkManager: this.initData?.networkManager,
-    });
+  setupSoloGame() {
+    this.gameManager = new GameManager(this);
 
-    // Get players from mode
-    this.players = this.gameMode.createPlayers(this);
+    // Create players
+    this.players = [];
+    const playerInfo = this.gameManager.playerInfo;
 
-    // Always create network UI (will only be shown if mode emits connection events)
-    this.createNetworkIndicator();
+    for (let i = 0; i < 4; i++) {
+      const player = new Player(
+        this,
+        i,
+        playerInfo[i].name,
+        playerInfo[i].emoji,
+        playerInfo[i].isHuman,
+        playerInfo[i].isHuman
+          ? (cardData: CardData) => this.onHumanCardPlayed(cardData)
+          : undefined,
+      );
+      this.players.push(player);
 
-    // Setup unified event listeners (no mode branching)
+      // Listen for card play events from human player
+      // if (player.isHuman) {
+      //   player.hand.on("cardPlayed", (cardData: CardData) => {
+      //     this.onHumanCardPlayed(cardData);
+      //   });
+      // }
+    }
+
+    this.gameManager.setPlayers(this.players);
+    this.gameManager.setTrickArea(this.trickArea);
+
+    // Setup event listeners
     this.setupEventListeners();
 
     // Launch UI scene
     this.scene.launch("UIScene", {
-      gameMode: this.gameMode,
+      gameManager: this.gameManager,
       audioManager: this.audioManager,
+      isMultiplayer: false,
     });
 
     // Start the game
-    await this.gameMode.startGame();
+    this.gameManager.startGame();
+  }
+
+  setupMultiplayer() {
+    if (!this.networkManager) {
+      console.error("No network manager provided for multiplayer!");
+      this.scene.start("MenuScene");
+      return;
+    }
+
+    // Create network indicator
+    this.createNetworkIndicator();
+
+    // Create players from network state
+    this.players = [];
+    const networkPlayers = this.networkManager.getPlayers();
+    const localId = this.networkManager.playerId;
+
+    // Find local player's seat index
+    const localPlayer = networkPlayers.find((p) => p.id === localId);
+    const localSeatIndex = localPlayer ? localPlayer.seatIndex : 0;
+
+    // Create player objects with relative positioning
+    // Local player should always be at position 0 (bottom)
+    networkPlayers.forEach((netPlayer) => {
+      const isLocal = netPlayer.id === localId;
+
+      // Calculate relative position: local player at 0 (bottom), others clockwise
+      const relativePosition = (netPlayer.seatIndex - localSeatIndex + 4) % 4;
+
+      const player = new Player(
+        this,
+        relativePosition,
+        netPlayer.name,
+        netPlayer.emoji,
+        isLocal, // isHuman = isLocal in multiplayer
+        isLocal
+          ? (cardData: CardData) => this.onMultiplayerCardPlayed(cardData)
+          : undefined,
+      );
+      this.players.push(player);
+
+      // Store network ID and absolute seat index for server communication
+      player.networkId = netPlayer.id;
+      player.absoluteSeatIndex = netPlayer.seatIndex;
+
+      // Listen for card play events from local player
+      // if (isLocal) {
+      //   player.hand.on("cardPlayed", (cardData: CardData) => {
+      //     this.onMultiplayerCardPlayed(cardData);
+      //   });
+      // }
+    });
+
+    // Sort players by relative position for consistent array indexing
+    // This ensures players[0] is always local player (bottom)
+    this.players.sort((a, b) => a.index - b.index);
+
+    // Setup network event listeners
+    this.setupNetworkListeners();
+
+    // Launch UI scene
+    this.scene.launch("UIScene", {
+      gameManager: null, // No local game manager in multiplayer
+      audioManager: this.audioManager,
+      isMultiplayer: true,
+      networkManager: this.networkManager,
+    });
+
+    // Deal initial hand if already dealt
+    this.syncHandFromServer();
   }
 
   createNetworkIndicator() {
     const { width } = this.cameras.main;
 
     // Create network indicator in top-right corner, to the left of settings icon
-    // Hidden by default, will be shown when connection events are received
+    // Settings icon is at (width - margin), so place this at (width - margin - 60)
     this.networkIndicator = new NetworkIndicator(this, width - 150, 50);
-    this.networkIndicator.container.setVisible(false);
 
     // Create reconnection overlay (hidden by default)
     this.reconnectionOverlay = new ReconnectionOverlay(this);
   }
 
-  setupConnectionListeners() {
-    // Listen to connection events from game mode (only multiplayer emits these)
-
+  setupNetworkListeners() {
     // Connection quality changes
-    this.gameMode.on(EVENTS.CONNECTION_QUALITY_CHANGED, ({ quality }: any) => {
+    this.networkManager.on("connectionQualityChange", ({ quality }: any) => {
       if (this.networkIndicator) {
-        this.networkIndicator.container.setVisible(true);
         this.networkIndicator.updateQuality(quality);
       }
     });
 
     // Reconnecting
-    this.gameMode.on(EVENTS.RECONNECTING, ({ attempt }: any) => {
+    this.networkManager.on("reconnecting", ({ attempt }: any) => {
       console.log("GameScene: Reconnecting...", attempt);
       if (this.networkIndicator) {
-        this.networkIndicator.container.setVisible(true);
         this.networkIndicator.showReconnecting(attempt);
       }
-      this.reconnectionOverlay?.show(attempt);
+      this.reconnectionOverlay.show(attempt);
     });
 
     // Reconnected
-    this.gameMode.on(EVENTS.RECONNECTED, ({ message }: any) => {
+    this.networkManager.on("reconnected", ({ message }: any) => {
       console.log("GameScene: Reconnected!", message);
       if (this.networkIndicator) {
         this.networkIndicator.showReconnected();
       }
-      this.reconnectionOverlay?.hide();
+      this.reconnectionOverlay.hide();
+
+      // Wait a brief moment for server state to fully sync, then re-sync game state
+      this.time.delayedCall(200, () => {
+        this.syncHandFromServer();
+      });
 
       // Show brief success message
       const successText = this.add
@@ -143,9 +233,9 @@ export default class GameScene extends Phaser.Scene {
     });
 
     // Reconnection failed
-    this.gameMode.on(EVENTS.RECONNECTION_FAILED, ({ message }: any) => {
+    this.networkManager.on("reconnectionFailed", ({ message }: any) => {
       console.log("GameScene: Reconnection failed", message);
-      this.reconnectionOverlay?.hide();
+      this.reconnectionOverlay.hide();
 
       // Show error and redirect to menu
       this.events.emit("connectionLost", { message });
@@ -156,25 +246,208 @@ export default class GameScene extends Phaser.Scene {
       });
     });
 
-    // Error handling
-    this.gameMode.on(EVENTS.CONNECTION_ERROR, ({ message }: any) => {
-      console.error("Network error:", message);
-      this.events.emit("networkError", {
-        message: `Connection error: ${message || "Unknown error"}`,
+    // Phase change
+    this.networkManager.on("phaseChange", ({ phase }: any) => {
+      this.events.emit("phaseChanged", phase);
+
+      if (phase === "roundEnd") {
+        const players = this.networkManager.getPlayers();
+        this.events.emit("roundComplete", {
+          players: players.map((p) => ({
+            name: p.name,
+            emoji: p.emoji,
+            bid: p.bid,
+            tricksWon: p.tricksWon,
+            roundScore: p.roundScore,
+            totalScore: p.score,
+          })),
+        });
+      }
+
+      if (phase === "gameOver") {
+        const players = this.networkManager.getPlayers();
+        const sortedPlayers = [...players].sort((a, b) => b.score - a.score);
+        const winner = sortedPlayers[0];
+        this.audioManager.playWinSound();
+        this.events.emit("gameComplete", {
+          winner: {
+            name: winner.name,
+            emoji: winner.emoji,
+            score: winner.score,
+          },
+          players: sortedPlayers.map((p) => ({
+            name: p.name,
+            emoji: p.emoji,
+            score: p.score,
+          })),
+        });
+      }
+    });
+
+    // Turn change
+    this.networkManager.on("turnChange", ({ playerId, isMyTurn }: any) => {
+      this.players.forEach((p) => {
+        if (p.networkId === playerId) {
+          p.showTurnIndicator();
+          if (isMyTurn) {
+            // Enable card selection for local player with proper validation
+            const leadSuit = this.networkManager.getLeadSuit();
+            const currentTrick = this.networkManager.getCurrentTrick();
+            const phase = this.networkManager.getPhase();
+            if (phase === "playing") {
+              p.hand.updatePlayableCards(leadSuit, currentTrick);
+            }
+          }
+        } else {
+          p.hideTurnIndicator();
+          p.hand.disableAllCards();
+        }
       });
     });
 
-    // Room left
-    this.gameMode.on(EVENTS.ROOM_LEFT, ({ code }: any) => {
-      console.log("Room left event received:", code);
+    // Card added to hand
+    this.networkManager.on("cardAdded", ({ card }: any) => {
+      const localPlayer = this.players.find(
+        (p) => p.networkId === this.networkManager.playerId,
+      );
+      if (localPlayer) {
+        localPlayer.hand.addCard(card);
+      }
+    });
 
+    // Card played by any player
+    this.networkManager.on("cardPlayed", ({ playerId, card }: any) => {
+      this.audioManager.playCardSound();
+
+      // Find player in local array
+      let player = this.players.find((p) => p.networkId === playerId);
+      let relativePosition;
+      let removedCard = null;
+
+      if (player) {
+        // Use relative position for visual display
+        relativePosition = player.index;
+
+        // Remove card from player's hand for animation
+        if (player.networkId === this.networkManager.playerId) {
+          // Local player - remove the actual card
+          removedCard = player.hand.removeCard(card.id);
+        } else {
+          // Remote player - remove any placeholder card for animation
+          removedCard = player.hand.removeFirstCard();
+        }
+      } else {
+        // Player not in local array (shouldn't happen in multiplayer, but fallback)
+        console.warn(`Player ${playerId} not found in local players array`);
+        relativePosition = 0;
+      }
+
+      // Add card to trick area using relative position for visual placement
+      this.trickArea.playCard(card, relativePosition, removedCard);
+    });
+
+    // Trick cleared
+    this.networkManager.on("trickCleared", () => {
+      this.trickArea.clear();
+    });
+
+    // Trick winner
+    this.networkManager.on("trickWinner", (winnerId: any) => {
+      const winner = this.players.find((p) => p.networkId === winnerId);
+      if (winner) {
+        // Update tricks won count
+        winner.addTrick();
+
+        if (winner.nameLabel) {
+          this.tweens.add({
+            targets: winner.nameLabel,
+            scaleX: 1.2,
+            scaleY: 1.2,
+            duration: 200,
+            yoyo: true,
+          });
+        }
+
+        // Animate cards being collected to the winner
+        const winnerIndex = this.players.indexOf(winner);
+        this.time.delayedCall(1000, () => {
+          this.trickArea.collectTrick(winnerIndex, 400 as any);
+        });
+      }
+      // For bots not in players array, just collect the trick without animation
+    });
+
+    // Player bid
+    this.networkManager.on("playerBid", ({ playerId, bid }: any) => {
+      const player = this.players.find((p) => p.networkId === playerId);
+      if (player) {
+        // Update the local player's bid and stats display
+        player.setBid(bid);
+        const playerIndex = this.players.indexOf(player);
+        this.events.emit("bidPlaced", { playerIndex, bid });
+      } else {
+        // This shouldn't happen in multiplayer - all players should be in the array
+        console.warn(`Player ${playerId} not found for bid event`);
+      }
+    });
+
+    // Round change
+    this.networkManager.on("roundChange", () => {
+      // Reset all players for new round
+      this.players.forEach((player) => {
+        player.reset();
+      });
+    });
+
+    // Lead suit changed - update playable cards if it's our turn
+    this.networkManager.on("leadSuitChange", (leadSuit: Suit) => {
+      if (
+        this.networkManager.isMyTurn() &&
+        this.networkManager.getPhase() === "playing"
+      ) {
+        const localPlayer = this.players.find(
+          (p) => p.networkId === this.networkManager.playerId,
+        );
+        if (localPlayer) {
+          const currentTrick = this.networkManager.getCurrentTrick();
+          localPlayer.hand.updatePlayableCards(leadSuit, currentTrick);
+        }
+      }
+    });
+
+    // Error handling
+    this.networkManager.on("error", (data: any) => {
+      console.error("Network error:", data);
+      // Show error message to user instead of silently redirecting
+      this.events.emit("networkError", {
+        message: `Connection error: ${data.message || "Unknown error"}`,
+      });
+    });
+
+    // Remote player hand changes (for visual card backs)
+    this.networkManager.on(
+      "remoteHandChanged",
+      ({ playerId, handCount }: any) => {
+        const player = this.players.find((p) => p.networkId === playerId);
+        if (player) {
+          player.hand.updateCardCount(handCount);
+        }
+      },
+    );
+
+    // Room left
+    this.networkManager.on("roomLeft", (data: any) => {
+      console.log("Room left event received:", data);
+
+      // Show a message before redirecting
       const message =
-        code === 1000
+        data?.code === 1000
           ? "Disconnected from game"
           : "Connection lost - returning to menu";
 
       this.events.emit("connectionLost", { message });
 
+      // Delay redirect to allow user to see what happened
       this.time.delayedCall(1500, () => {
         this.scene.stop("UIScene");
         this.scene.start("MenuScene");
@@ -182,11 +455,78 @@ export default class GameScene extends Phaser.Scene {
     });
   }
 
-  setupEventListeners() {
-    // Unified event listeners - no mode conditionals!
+  syncHandFromServer() {
+    // Get current hand from server and display
+    const hand = this.networkManager.getMyHand();
+    const localPlayer = this.players.find(
+      (p) => p.networkId === this.networkManager.playerId,
+    );
 
-    // Turn changed - now both modes emit playerIndex consistently
-    this.gameMode.on(EVENTS.TURN_CHANGED, ({ playerIndex, isMyTurn }: any) => {
+    if (localPlayer) {
+      // Always sync cards, even if hand is empty (cards might have been played)
+      console.log("GameScene: Syncing hand from server. Cards:", hand.length);
+      localPlayer.hand.setCards(hand, false);
+    }
+
+    // Sync card counts for all remote players (show card backs)
+    const state = this.networkManager.getState();
+    if (state) {
+      state.players.forEach((player: any, sessionId: string) => {
+        if (sessionId !== this.networkManager.playerId) {
+          const handCount = player.hand.length;
+          const localPlayerObj = this.players.find(
+            (p) => p.networkId === sessionId,
+          );
+          if (localPlayerObj) {
+            console.log(
+              `GameScene: Updating card count for ${player.name}: ${handCount}`,
+            );
+            localPlayerObj.hand.updateCardCount(handCount);
+          }
+        }
+      });
+
+      // Check current game state and update UI accordingly
+      const phase = state.phase;
+      const currentTurnPlayerId = state.currentTurn;
+      const isMyTurn = currentTurnPlayerId === this.networkManager.playerId;
+
+      // Update turn indicators for all players
+      this.players.forEach((p) => {
+        if (p.networkId === currentTurnPlayerId) {
+          p.showTurnIndicator();
+        } else {
+          p.hideTurnIndicator();
+          p.hand.disableAllCards();
+        }
+      });
+
+      if (phase === "bidding" && isMyTurn) {
+        // Emit phase changed event to trigger bidding UI
+        this.time.delayedCall(500, () => {
+          this.events.emit("phaseChanged", phase);
+        });
+      } else if (phase === "playing" && isMyTurn && localPlayer) {
+        // If it's our turn to play, update playable cards
+        const leadSuit = state.leadSuit || "";
+        const currentTrick = this.networkManager.getCurrentTrick();
+        console.log(
+          "GameScene: Reconnected during our turn, updating playable cards. Lead suit:",
+          leadSuit,
+        );
+        localPlayer.hand.updatePlayableCards(leadSuit, currentTrick);
+      }
+    }
+  }
+
+  onMultiplayerCardPlayed(cardData: CardData) {
+    // Send card play to server
+    this.networkManager.sendPlayCard(cardData.id);
+  }
+
+  setupEventListeners() {
+    // Turn changed
+    this.gameManager.on(EVENTS.TURN_CHANGED, (playerIndex: number) => {
       this.players.forEach((p, i) => {
         if (i === playerIndex) {
           p.showTurnIndicator();
@@ -196,61 +536,100 @@ export default class GameScene extends Phaser.Scene {
       });
     });
 
+    // Phase changed
+    this.gameManager.on(EVENTS.PHASE_CHANGED, (phase: any) => {
+      this.events.emit("phaseChanged", phase);
+    });
+
     // Card played
-    this.gameMode.on(EVENTS.CARD_PLAYED, () => {
+    this.gameManager.on(EVENTS.CARD_PLAYED, () => {
       this.audioManager.playCardSound();
     });
 
     // Trick complete
-    this.gameMode.on(EVENTS.TRICK_COMPLETE, ({ winnerIndex }: any) => {
+    this.gameManager.on(EVENTS.TRICK_COMPLETE, ({ winnerIndex }: any) => {
+      // Flash winner indicator
       const winner = this.players[winnerIndex];
-      if (winner && winner.nameLabel) {
-        this.tweens.add({
-          targets: winner.nameLabel,
-          scaleX: 1.2,
-          scaleY: 1.2,
-          duration: 200,
-          yoyo: true,
-        });
-      }
+      this.tweens.add({
+        targets: winner.nameLabel,
+        scaleX: 1.2,
+        scaleY: 1.2,
+        duration: 200,
+        yoyo: true,
+      });
     });
 
-    // Game complete - play win sound
-    this.gameMode.on(EVENTS.GAME_COMPLETE, () => {
+    // Round complete
+    this.gameManager.on(EVENTS.ROUND_COMPLETE, (data: any) => {
+      this.events.emit("roundComplete", data);
+    });
+
+    // Game complete
+    this.gameManager.on(EVENTS.GAME_COMPLETE, (data: any) => {
       this.audioManager.playWinSound();
+      this.events.emit("gameComplete", data);
     });
 
-    // Setup connection listeners (only fired by multiplayer mode)
-    this.setupConnectionListeners();
+    // Bid placed
+    this.gameManager.on(EVENTS.BID_PLACED, ({ playerIndex, bid }: any) => {
+      this.events.emit("bidPlaced", { playerIndex, bid });
+    });
+  }
+
+  onHumanCardPlayed(cardData: CardData) {
+    this.gameManager.playCard(cardData, 0);
+  }
+
+  // Called from UIScene when human places bid
+  onHumanBid(bid: number) {
+    this.gameManager.placeHumanBid(bid);
   }
 
   // Called from UIScene to continue to next round
   continueToNextRound() {
-    this.gameMode.continueToNextRound();
+    if (this.isMultiplayer) {
+      this.networkManager.sendNextRound();
+    } else {
+      this.gameManager.continueToNextRound();
+    }
   }
 
   // Called from UIScene to restart game
   restartGame() {
     this.trickArea.clear();
-    this.gameMode.restartGame();
+    if (this.isMultiplayer && this.networkManager) {
+      this.networkManager.sendRestart();
+    } else if (this.gameManager) {
+      this.gameManager.restartGame();
+    }
+  }
+
+  // Called from UIScene when human places bid (multiplayer)
+  onMultiplayerBid(bid: number) {
+    if (this.networkManager) {
+      this.networkManager.sendBid(bid);
+    }
   }
 
   // Called from UIScene to return to menu
-  async returnToMenu() {
+  returnToMenu() {
     this.audioManager.destroy();
+
+    // Clean up multiplayer connection
+    if (this.isMultiplayer && this.networkManager) {
+      (this.networkManager as any).removeAllListeners();
+      this.networkManager.leaveRoom();
+    }
 
     // Clean up network indicator
     if (this.networkIndicator) {
-      this.networkIndicator.destroy();
+      (this.networkIndicator as any).destroy();
     }
 
     // Clean up reconnection overlay
     if (this.reconnectionOverlay) {
       this.reconnectionOverlay.destroy();
     }
-
-    // Cleanup game mode (this will handle network cleanup for multiplayer)
-    await this.gameMode.cleanup();
 
     this.scene.stop("UIScene");
     this.scene.start("MenuScene");
