+++ src/scenes/UIScene.ts
@@ -1,19 +1,22 @@
 import Phaser from "phaser";
+import { EVENTS, PHASE } from "../utils/constants";
 import ScoreBoard from "../objects/game/ScoreBoard";
 import BiddingUI from "../objects/game/BiddingUI";
 import RoundModal from "../objects/game/RoundModal";
 import GameOverModal from "../objects/game/GameOverModal";
 import SettingsModal from "../objects/game/SettingsModal";
+import NetworkManager from "../managers/NetworkManager";
+import GameManager from "../managers/GameManager";
 import AudioManager from "../managers/AudioManager";
 import GameScene from "./GameScene";
 import Common from "../objects/game/Common";
-import type { GameModeBase } from "../modes/GameModeBase";
-import { EVENTS, UI_TIMING } from "../utils/constants";
 
 export default class UIScene extends Phaser.Scene {
-  private gameMode!: GameModeBase;
-  private audioManager!: AudioManager;
-  private gameScene!: GameScene;
+  isMultiplayer: boolean;
+  networkManager: NetworkManager | null;
+  gameManager!: GameManager;
+  audioManager!: AudioManager;
+  gameScene!: GameScene;
   scoreBoard!: ScoreBoard;
   roundModal!: RoundModal;
   gameOverModal!: GameOverModal;
@@ -22,23 +25,26 @@ export default class UIScene extends Phaser.Scene {
 
   constructor() {
     super({ key: "UIScene" });
+    this.isMultiplayer = false;
+    this.networkManager = null;
   }
 
   init(data: any) {
-    this.gameMode = data.gameMode;
+    this.gameManager = data.gameManager;
     this.audioManager = data.audioManager;
+    this.isMultiplayer = data.isMultiplayer || false;
+    this.networkManager = data.networkManager || null;
   }
 
   create() {
     // Get reference to game scene
     this.gameScene = this.scene.get("GameScene") as GameScene;
 
-    // Create scoreboard - unified for both modes!
     this.scoreBoard = new ScoreBoard(
       this,
-      false, // We'll update this based on game mode later if needed
-      this.gameMode.getPlayers(),
-      this.gameMode.getCurrentRound(),
+      this.isMultiplayer,
+      this.getPlayersData(),
+      this.getCurrentRound(),
     );
 
     // Create modals
@@ -64,74 +70,80 @@ export default class UIScene extends Phaser.Scene {
     this.settingsModal = new SettingsModal(this, {
       audioManager: this.audioManager,
       onQuit: () => this.gameScene.returnToMenu(),
-      onNewGame: this.isMultiplayer()
-        ? null
-        : () => this.gameScene.restartGame(),
+      onNewGame: !this.isMultiplayer
+        ? () => this.gameScene.restartGame()
+        : null,
     });
 
     // Create bidding UI
     this.biddingUI = new BiddingUI(
       this,
-      (bid) => this.gameMode.onBidSelected(bid),
+      (bid) => this.onBidSelected(bid),
       this.audioManager,
     );
 
-    // Setup unified event listeners
+    // Listen for game events
     this.setupEventListeners();
   }
 
+  getPlayersData() {
+    if (this.isMultiplayer && this.networkManager) {
+      return this.networkManager.getPlayers();
+    } else if (this.gameManager) {
+      return this.gameManager.getPlayers();
+    }
+    return [];
+  }
+
+  getCurrentRound() {
+    if (this.isMultiplayer && this.networkManager) {
+      const state = this.networkManager.getState();
+      return state?.currentRound || 1;
+    } else if (this.gameManager) {
+      return this.gameManager.getCurrentRound();
+    }
+    return 1;
+  }
+
+  getPhase() {
+    if (this.isMultiplayer && this.networkManager) {
+      return this.networkManager.getPhase();
+    } else if (this.gameManager) {
+      return this.gameManager.getPhase();
+    }
+    return "waiting";
+  }
+
+  onBidSelected(bid: number) {
+    if (this.isMultiplayer) {
+      this.gameScene.onMultiplayerBid(bid);
+    } else {
+      this.gameScene.onHumanBid(bid);
+    }
+  }
+
   setupEventListeners() {
-    // Unified event listeners - no mode conditionals!
+    if (this.isMultiplayer) {
+      this.setupMultiplayerEventListeners();
+    } else {
+      this.setupSoloEventListeners();
+    }
 
-    // Phase changed
-    this.gameMode.on(EVENTS.PHASE_CHANGED, (phase: string) => {
-      // Update scoreboard
-      this.scoreBoard.updateScoreboard(
-        this.gameMode.getPlayers(),
-        this.gameMode.getCurrentRound(),
-      );
-
-      // Note: Don't show bidding UI here!
-      // It will be shown by the turnChanged event when it's actually the player's turn
-    });
-
-    // Turn changed
-    this.gameMode.on(EVENTS.TURN_CHANGED, ({ isMyTurn }: any) => {
-      const phase = this.gameMode.getPhase();
-      if (phase === "bidding" && isMyTurn) {
-        // Small delay to allow card animations to settle before showing UI
-        this.time.delayedCall(UI_TIMING.BIDDING_UI_DELAY, () => this.biddingUI.show());
-      }
-    });
-
-    // Bid placed
-    this.gameMode.on(EVENTS.BID_PLACED, ({ playerIndex }: any) => {
-      // Update scoreboard
-      this.scoreBoard.updateScoreboard(
-        this.gameMode.getPlayers(),
-        this.gameMode.getCurrentRound(),
-      );
-
-      // Hide bidding UI if it was the local player
-      if (this.gameMode.isLocalPlayer(playerIndex)) {
-        this.biddingUI.hide();
-      }
-    });
-
+    // Common event listeners
     // Round complete
-    this.gameMode.on(EVENTS.ROUND_COMPLETE, (data: any) => {
+    this.gameScene.events.on("roundComplete", (data: any) => {
       this.scoreBoard.updateScoreboard(
-        this.gameMode.getPlayers(),
-        this.gameMode.getCurrentRound(),
+        this.getPlayersData(),
+        this.getCurrentRound(),
       );
       this.time.delayedCall(500, () => this.roundModal.showRoundResults(data));
     });
 
     // Game complete
-    this.gameMode.on(EVENTS.GAME_COMPLETE, (data: any) => {
+    this.gameScene.events.on("gameComplete", (data: any) => {
       this.scoreBoard.updateScoreboard(
-        this.gameMode.getPlayers(),
-        this.gameMode.getCurrentRound(),
+        this.getPlayersData(),
+        this.getCurrentRound(),
       );
       this.time.delayedCall(500, () =>
         this.gameOverModal.showGameResults(data),
@@ -139,9 +151,73 @@ export default class UIScene extends Phaser.Scene {
     });
   }
 
-  private isMultiplayer(): boolean {
-    // Check if any player has an id (multiplayer players have network IDs)
-    const players = this.gameMode.getPlayers();
-    return players.some((p) => p.id !== undefined);
+  setupSoloEventListeners() {
+    // Listen for turn changes during bidding
+    this.gameManager?.on(EVENTS.TURN_CHANGED, (playerIndex: number) => {
+      const phase = this.gameManager?.phase;
+      if (phase === PHASE.BIDDING && playerIndex === 0) {
+        // It's the human's turn to bid
+        this.time.delayedCall(300, () => this.biddingUI.show());
+      }
+    });
+
+    // Bid placed - no need to manually check for next bidder anymore
+    // The TURN_CHANGED event will handle showing the UI
+    this.gameScene.events.on("bidPlaced", ({ playerIndex }: any) => {
+      // Just hide if it was the human who bid
+      if (playerIndex === 0) {
+        this.biddingUI.hide();
+      }
+    });
+  }
+
+  setupMultiplayerEventListeners() {
+    // Listen for phase changes - show bidding UI when it's our turn
+    this.gameScene.events.on("phaseChanged", (phase: any) => {
+      this.scoreBoard.updateScoreboard(
+        this.getPlayersData(),
+        this.getCurrentRound(),
+      );
+      if (phase === "bidding") {
+        // Add slight delay to ensure currentTurn state is updated
+        this.time.delayedCall(100, () => {
+          if (this.networkManager!.isMyTurn()) {
+            this.biddingUI.show();
+          }
+        });
+      }
+    });
+
+    // Turn change - show bidding UI if it's bidding phase and our turn
+    this.networkManager!.on("turnChange", ({ isMyTurn }: any) => {
+      const phase = this.networkManager!.getPhase();
+      if (phase === "bidding" && isMyTurn) {
+        this.time.delayedCall(300, () => this.biddingUI.show());
+      }
+    });
+
+    // Score updates
+    this.networkManager!.on("playerScoreChange", () => {
+      this.scoreBoard.updateScoreboard(
+        this.getPlayersData(),
+        this.getCurrentRound(),
+      );
+    });
+
+    // Bid placed
+    this.networkManager!.on("playerBid", () => {
+      this.scoreBoard.updateScoreboard(
+        this.getPlayersData(),
+        this.getCurrentRound(),
+      );
+    });
+
+    // Tricks won update
+    this.networkManager!.on("playerTricksWon", () => {
+      this.scoreBoard.updateScoreboard(
+        this.getPlayersData(),
+        this.getCurrentRound(),
+      );
+    });
   }
 }
