+++ src/utils/constants.ts
@@ -113,21 +113,9 @@ export const EVENTS = {
   TRICK_COMPLETE: "trickComplete",
   ROUND_COMPLETE: "roundComplete",
   GAME_COMPLETE: "gameComplete",
-  BID_PLACED: "bidPlayed",
+  BID_PLACED: "bidPlaced",
   TURN_CHANGED: "turnChanged",
   PHASE_CHANGED: "phaseChanged",
-  // Connection events (multiplayer only, ignored by solo mode)
-  CONNECTION_QUALITY_CHANGED: "connectionQualityChanged",
-  RECONNECTING: "reconnecting",
-  RECONNECTED: "reconnected",
-  RECONNECTION_FAILED: "reconnectionFailed",
-  CONNECTION_ERROR: "connectionError",
-  ROOM_LEFT: "roomLeft",
-} as const;
-
-// UI timing constants
-export const UI_TIMING = {
-  BIDDING_UI_DELAY: 300, // Delay before showing bidding UI to allow card animations to settle
 } as const;
 
 export type GameEvent = (typeof EVENTS)[keyof typeof EVENTS];
